change category endpoints (GET, POST, PUT, DELETE)
